using unsigned long long ints for nrgeo in igraph_{edge}_betweenness_estimate

+++ src/centrality.c
@@ -1061,7 +1061,7 @@ int igraph_betweenness_estimate(const igraph_t *graph, igraph_vector_t *res,
   long int no_of_nodes=igraph_vcount(graph);
   igraph_dqueue_t q=IGRAPH_DQUEUE_NULL;
   long int *distance;
-  long int *nrgeo;
+  unsigned long long int *nrgeo;  /* must be long long; consider grid graphs for example */
   double *tmpscore;
   igraph_stack_t stack=IGRAPH_STACK_NULL;
   long int source;
@@ -1117,7 +1117,7 @@ int igraph_betweenness_estimate(const igraph_t *graph, igraph_vector_t *res,
     IGRAPH_ERROR("betweenness failed", IGRAPH_ENOMEM);
   }
   IGRAPH_FINALLY(igraph_free, distance);
-  nrgeo=igraph_Calloc(no_of_nodes, long int);
+  nrgeo=igraph_Calloc(no_of_nodes, unsigned long long int);
   if (nrgeo==0) {
     IGRAPH_ERROR("betweenness failed", IGRAPH_ENOMEM);
   }
@@ -1460,7 +1460,7 @@ int igraph_edge_betweenness_estimate(const igraph_t *graph, igraph_vector_t *res
   long int no_of_edges=igraph_ecount(graph);
   igraph_dqueue_t q=IGRAPH_DQUEUE_NULL;
   long int *distance;
-  long int *nrgeo;
+  unsigned long long int *nrgeo;
   double *tmpscore;
   igraph_stack_t stack=IGRAPH_STACK_NULL;
   long int source;
@@ -1500,7 +1500,7 @@ int igraph_edge_betweenness_estimate(const igraph_t *graph, igraph_vector_t *res
     IGRAPH_ERROR("edge betweenness failed", IGRAPH_ENOMEM);
   }
   IGRAPH_FINALLY(igraph_free, distance);
-  nrgeo=igraph_Calloc(no_of_nodes, long int);
+  nrgeo=igraph_Calloc(no_of_nodes, unsigned long long int);
   if (nrgeo==0) {
     IGRAPH_ERROR("edge betweenness failed", IGRAPH_ENOMEM);
   }
@@ -1526,7 +1526,7 @@ int igraph_edge_betweenness_estimate(const igraph_t *graph, igraph_vector_t *res
     IGRAPH_ALLOW_INTERRUPTION();
 
     memset(distance, 0, no_of_nodes*sizeof(long int));
-    memset(nrgeo, 0, no_of_nodes*sizeof(long int));
+    memset(nrgeo, 0, no_of_nodes*sizeof(unsigned long long int));
     memset(tmpscore, 0, no_of_nodes*sizeof(double));
     igraph_stack_clear(&stack); /* it should be empty anyway... */
